Remove useEffect errors
