+++ src/screens/TaskScreen.js
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React from "react";
 import { Header, SideBar, Body } from "../components/";
 import {
   Container,
